Add BenchmarkCLI executable to BenchmarkCore package
- Moves CLI functionality into BenchmarkCore package
- Eliminates separate TokenTraceCLI package
- Uses Apple's XMLParser instead of XMLCoder dependency
- Adds BenchmarkCLI executable target
- Includes run-trace.sh script for xctrace workflow
- Provides parse-xml command for XML parsing
