+++ .vscode/launch.json
@@ -1,0 +1,24 @@
+{
+    "configurations": [
+        {
+            "type": "swift",
+            "request": "launch",
+            "args": [],
+            "cwd": "${workspaceFolder:Foundation-Models-Framework-Benchmark}/BenchmarkCore",
+            "name": "Debug BenchmarkCLI (BenchmarkCore)",
+            "target": "BenchmarkCLI",
+            "configuration": "debug",
+            "preLaunchTask": "swift: Build Debug BenchmarkCLI (BenchmarkCore)"
+        },
+        {
+            "type": "swift",
+            "request": "launch",
+            "args": [],
+            "cwd": "${workspaceFolder:Foundation-Models-Framework-Benchmark}/BenchmarkCore",
+            "name": "Release BenchmarkCLI (BenchmarkCore)",
+            "target": "BenchmarkCLI",
+            "configuration": "release",
+            "preLaunchTask": "swift: Build Release BenchmarkCLI (BenchmarkCore)"
+        }
+    ]
+}
