+++ FoundationStudio/FoundationStudioTests/FoundationStudioTests.swift
@@ -1,0 +1,16 @@
+//
+//  FoundationStudioTests.swift
+//  FoundationStudioTests
+//
+//  Created by Rudrank Riyam on 12/1/25.
+//
+
+import Testing
+
+struct FoundationStudioTests {
+
+    @Test func example() async throws {
+        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+    }
+
+}
