--- TokenTraceCLI/Package.resolved
@@ -1,15 +1,0 @@
-{
-  "originHash" : "6a91eb86d8a2e84828719c2c000ba8404004213eadc2b43851e234442748cd67",
-  "pins" : [
-    {
-      "identity" : "xmlcoder",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/MaxDesiatov/XMLCoder.git",
-      "state" : {
-        "revision" : "b1e944cbd0ef33787b13f639a5418d55b3bed501",
-        "version" : "0.17.1"
-      }
-    }
-  ],
-  "version" : 3
-}
